+++ pymc/distributions/discrete.py
@@ -48,7 +48,7 @@ from pymc.distributions.dist_math import (
 from pymc.distributions.distribution import Discrete
 from pymc.distributions.mixture import Mixture
 from pymc.distributions.shape_utils import rv_size_is_none
-from pymc.logprob.joint_logprob import logp
+from pymc.logprob.basic import logp
 from pymc.math import sigmoid
 from pymc.pytensorf import floatX, intX
 from pymc.vartypes import continuous_types
