+++ pymc/logprob/basic.py
